backend/pyproject.toml
[project]
name = "finance-backend"
version = "0.1.0"
description = "FinanceBoard API"
authors = [
    {name = "Engineering Candidate"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (==0.115.12)",
    "uvicorn[standard] (>=0.43.0,<0.44.0)",
    "sqlalchemy (==2.0.40)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "bcrypt (>=4.0.1,<5.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
